chore: rename dev/build to dev/internal/build
