+++ dev/sdk_typescript.go
@@ -8,7 +8,7 @@ import (
 
 	"golang.org/x/sync/errgroup"
 
-	"github.com/dagger/dagger/dev/build"
+	"github.com/dagger/dagger/dev/internal/build"
 	"github.com/dagger/dagger/dev/internal/dagger"
 	"github.com/dagger/dagger/dev/util"
 )
